Add class picker and SRD info for Cleric
Change the functionality on the "Add Class" button to become a dropdown.
The first option in the dropdown is unsurprisingly named "Add Class",
which opens a dialog that lists the twelve SRD classes. Choosing one of
these radio buttons and clicking "Add" will add a new class with that
name at Level 1. Additionally, if the user chooses "Cleric", some
information (effects and proficiencies) will be pre-filled. Prefilling
for other classes will come soon.

The second option in the dropdown is "Create New..." and is the old
functionality that allows for flexibility of custom classes. Note that
creating a class in this way named e.g. "Cleric" will not prepopulate
the information above, which I think is the proper behavior.